pyproject.toml
# Setup information - see below for dev tool configuration
[build-system]
requires = ["flit_core >=3.4,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "soft7"
authors = [
    {name = "Thomas Fjæstad Hagelien", email = "thomas.f.hagelien@sintef.no"},
    {name = "Casper Welzel Andersen", email = "casper.w.andersen@sintef.no"},
    {name = "Treesa Rose Joseph", email = "treesa.joseph@sintef.no"},
]
description = "SOFT7 semantic interoperability framework."
readme = "README.md"
license = {file = "LICENSE"}
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Natural Language :: English",
    "Operating System :: OS Independent",
]
keywords = ["interoperability", "semantic"]
requires-python = "~=3.9"
dynamic = ["version"]

dependencies = [
    "graphviz",
    "httpx~=0.25.1",
    "Jinja2",
    "oteapi-core>=0.6.0,<1",
    "otelib>=0.4.0,<1",
    "pydantic~=2.4",
    "pymongo",
    "pyyaml",
    "rdflib",
]

[project.optional-dependencies]
testing = [
    "pytest~=7.4",
    "pytest-cov~=4.1",
    "requests-mock~=1.11",
]
dev = [
    "pre-commit~=2.18",
    "soft7[testing]",
]

[project.urls]
Home = "https://github.com/SINTEF/soft7"
Documentation = "https://SINTEF.github.io/soft7"
Source = "https://github.com/SINTEF/soft7"
"Issue Tracker" = "https://github.com/SINTEF/soft7/issues"
Changelog = "https://github.com/SINTEF/soft7/blob/main/CHANGELOG.md"
Package = "https://pypi.org/project/soft7"

[tool.flit.module]
name = "s7"

# Dev tool configurations
[tool.mypy]
python_version = "3.9"
ignore_missing_imports = true
scripts_are_modules = true
warn_unused_configs = true
show_error_codes = true
allow_redefinition = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
minversion = "7.4"
addopts = "-rs --cov=s7 --cov-report=term-missing"
filterwarnings = [
    "error",  # Fail on any warning
]

[tool.ruff.lint]
select = [
  # pycodestyle
  "E",
  # Pyflakes
  "F",
  # pyupgrade
  "UP",
  # flake8-bugbear
  "B",
  # flake8-simplify
  "SIM",
  # isort
  "I",
  # ruff
  "RUF",
]
